add taskType method
