+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/InternalProcessorContext.java
@@ -19,6 +19,9 @@ package org.apache.kafka.streams.processor.internals;
 import org.apache.kafka.streams.processor.ProcessorContext;
 import org.apache.kafka.streams.processor.RecordContext;
 import org.apache.kafka.streams.processor.StateStore;
+import org.apache.kafka.streams.processor.internals.ProcessorNode;
+import org.apache.kafka.streams.processor.internals.ProcessorRecordContext;
+import org.apache.kafka.streams.processor.internals.Task.TaskType;
 import org.apache.kafka.streams.processor.internals.metrics.StreamsMetricsImpl;
 import org.apache.kafka.streams.state.StoreBuilder;
 import org.apache.kafka.streams.state.internals.ThreadCache;
@@ -78,6 +81,11 @@ public interface InternalProcessorContext extends ProcessorContext {
      * Mark this context as being uninitialized
      */
     void uninitialize();
+
+    /**
+     * @return the type of task (active/standby/global) that this context corresponds to
+     */
+    TaskType taskType();
 
     /**
      * Get a correctly typed state store, given a handle on the original builder.
