+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/ProcessorContextImpl.java
@@ -16,49 +16,41 @@
  */
 package org.apache.kafka.streams.processor.internals;
 
-import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.StreamsConfig;
 import org.apache.kafka.streams.errors.StreamsException;
 import org.apache.kafka.streams.internals.ApiUtils;
-import org.apache.kafka.streams.kstream.Windowed;
 import org.apache.kafka.streams.processor.Cancellable;
-import org.apache.kafka.streams.processor.ProcessorContext;
 import org.apache.kafka.streams.processor.PunctuationType;
 import org.apache.kafka.streams.processor.Punctuator;
 import org.apache.kafka.streams.processor.StateStore;
 import org.apache.kafka.streams.processor.TaskId;
 import org.apache.kafka.streams.processor.To;
+import org.apache.kafka.streams.processor.internals.RecordCollector.Supplier;
+import org.apache.kafka.streams.processor.internals.Task.TaskType;
 import org.apache.kafka.streams.processor.internals.metrics.StreamsMetricsImpl;
-import org.apache.kafka.streams.state.KeyValueIterator;
-import org.apache.kafka.streams.state.KeyValueStore;
-import org.apache.kafka.streams.state.SessionStore;
-import org.apache.kafka.streams.state.TimestampedKeyValueStore;
-import org.apache.kafka.streams.state.TimestampedWindowStore;
-import org.apache.kafka.streams.state.ValueAndTimestamp;
-import org.apache.kafka.streams.state.WindowStore;
-import org.apache.kafka.streams.state.WindowStoreIterator;
 import org.apache.kafka.streams.state.internals.ThreadCache;
-import org.apache.kafka.streams.state.internals.WrappedStateStore;
 
 import java.time.Duration;
 import java.util.List;
 
 import static org.apache.kafka.streams.internals.ApiUtils.prepareMillisCheckFailMsgPrefix;
+import static org.apache.kafka.streams.processor.internals.AbstractReadOnlyDecorator.getReadOnlyStore;
+import static org.apache.kafka.streams.processor.internals.AbstractReadWriteDecorator.getReadWriteStore;
 
-public class ProcessorContextImpl extends AbstractProcessorContext implements RecordCollector.Supplier {
+public class ProcessorContextImpl extends AbstractProcessorContext implements Supplier {
 
     private final StreamTask task;
     private final RecordCollector collector;
     private final ToInternal toInternal = new ToInternal();
     private final static To SEND_TO_ALL = To.all();
 
-    ProcessorContextImpl(final TaskId id,
-                         final StreamTask task,
-                         final StreamsConfig config,
-                         final RecordCollector collector,
-                         final ProcessorStateManager stateMgr,
-                         final StreamsMetricsImpl metrics,
-                         final ThreadCache cache) {
+    public ProcessorContextImpl(final TaskId id,
+                                final StreamTask task,
+                                final StreamsConfig config,
+                                final RecordCollector collector,
+                                final ProcessorStateManager stateMgr,
+                                final StreamsMetricsImpl metrics,
+                                final ThreadCache cache) {
         super(id, config, metrics, stateMgr, cache);
         this.task = task;
         this.collector = collector;
@@ -82,21 +74,9 @@ public class ProcessorContextImpl extends AbstractProcessorContext implements Re
             throw new StreamsException("Accessing from an unknown node");
         }
 
-        final StateStore global = stateManager.getGlobalStore(name);
-        if (global != null) {
-            if (global instanceof TimestampedKeyValueStore) {
-                return new TimestampedKeyValueStoreReadOnlyDecorator<>((TimestampedKeyValueStore<?, ?>) global);
-            } else if (global instanceof KeyValueStore) {
-                return new KeyValueStoreReadOnlyDecorator<>((KeyValueStore<?, ?>) global);
-            } else if (global instanceof TimestampedWindowStore) {
-                return new TimestampedWindowStoreReadOnlyDecorator<>((TimestampedWindowStore<?, ?>) global);
-            } else if (global instanceof WindowStore) {
-                return new WindowStoreReadOnlyDecorator<>((WindowStore<?, ?>) global);
-            } else if (global instanceof SessionStore) {
-                return new SessionStoreReadOnlyDecorator<>((SessionStore<?, ?>) global);
-            }
-
-            return global;
+        final StateStore globalStore = stateManager.getGlobalStore(name);
+        if (globalStore != null) {
+            return getReadOnlyStore(globalStore);
         }
 
         if (!currentNode().stateStores.contains(name)) {
@@ -110,19 +90,7 @@ public class ProcessorContextImpl extends AbstractProcessorContext implements Re
         }
 
         final StateStore store = stateManager.getStore(name);
-        if (store instanceof TimestampedKeyValueStore) {
-            return new TimestampedKeyValueStoreReadWriteDecorator<>((TimestampedKeyValueStore<?, ?>) store);
-        } else if (store instanceof KeyValueStore) {
-            return new KeyValueStoreReadWriteDecorator<>((KeyValueStore<?, ?>) store);
-        } else if (store instanceof TimestampedWindowStore) {
-            return new TimestampedWindowStoreReadWriteDecorator<>((TimestampedWindowStore<?, ?>) store);
-        } else if (store instanceof WindowStore) {
-            return new WindowStoreReadWriteDecorator<>((WindowStore<?, ?>) store);
-        } else if (store instanceof SessionStore) {
-            return new SessionStoreReadWriteDecorator<>((SessionStore<?, ?>) store);
-        }
-
-        return store;
+        return getReadWriteStore(store);
     }
 
     @Override
@@ -221,410 +189,9 @@ public class ProcessorContextImpl extends AbstractProcessorContext implements Re
         return schedule(ApiUtils.validateMillisecondDuration(interval, msgPrefix), type, callback);
     }
 
-    private abstract static class StateStoreReadOnlyDecorator<T extends StateStore, K, V>
-        extends WrappedStateStore<T, K, V> {
-
-        static final String ERROR_MESSAGE = "Global store is read only";
-
-        private StateStoreReadOnlyDecorator(final T inner) {
-            super(inner);
-        }
-
-        @Override
-        public void flush() {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public void init(final ProcessorContext context,
-                         final StateStore root) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public void close() {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
+    @Override
+    public TaskType taskType() {
+        return stateManager.taskType();
     }
 
-    private static class KeyValueStoreReadOnlyDecorator<K, V>
-        extends StateStoreReadOnlyDecorator<KeyValueStore<K, V>, K, V>
-        implements KeyValueStore<K, V> {
-
-        private KeyValueStoreReadOnlyDecorator(final KeyValueStore<K, V> inner) {
-            super(inner);
-        }
-
-        @Override
-        public V get(final K key) {
-            return wrapped().get(key);
-        }
-
-        @Override
-        public KeyValueIterator<K, V> range(final K from,
-                                            final K to) {
-            return wrapped().range(from, to);
-        }
-
-        @Override
-        public KeyValueIterator<K, V> all() {
-            return wrapped().all();
-        }
-
-        @Override
-        public long approximateNumEntries() {
-            return wrapped().approximateNumEntries();
-        }
-
-        @Override
-        public void put(final K key,
-                        final V value) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public V putIfAbsent(final K key,
-                             final V value) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public void putAll(final List<KeyValue<K, V>> entries) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public V delete(final K key) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-    }
-
-    private static class TimestampedKeyValueStoreReadOnlyDecorator<K, V>
-        extends KeyValueStoreReadOnlyDecorator<K, ValueAndTimestamp<V>>
-        implements TimestampedKeyValueStore<K, V> {
-
-        private TimestampedKeyValueStoreReadOnlyDecorator(final TimestampedKeyValueStore<K, V> inner) {
-            super(inner);
-        }
-    }
-
-    private static class WindowStoreReadOnlyDecorator<K, V>
-        extends StateStoreReadOnlyDecorator<WindowStore<K, V>, K, V>
-        implements WindowStore<K, V> {
-
-        private WindowStoreReadOnlyDecorator(final WindowStore<K, V> inner) {
-            super(inner);
-        }
-
-        @Deprecated
-        @Override
-        public void put(final K key,
-                        final V value) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public void put(final K key,
-                        final V value,
-                        final long windowStartTimestamp) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public V fetch(final K key,
-                       final long time) {
-            return wrapped().fetch(key, time);
-        }
-
-        @Override
-        @Deprecated
-        public WindowStoreIterator<V> fetch(final K key,
-                                            final long timeFrom,
-                                            final long timeTo) {
-            return wrapped().fetch(key, timeFrom, timeTo);
-        }
-
-        @Override
-        @Deprecated
-        public KeyValueIterator<Windowed<K>, V> fetch(final K from,
-                                                      final K to,
-                                                      final long timeFrom,
-                                                      final long timeTo) {
-            return wrapped().fetch(from, to, timeFrom, timeTo);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, V> all() {
-            return wrapped().all();
-        }
-
-        @Override
-        @Deprecated
-        public KeyValueIterator<Windowed<K>, V> fetchAll(final long timeFrom,
-                                                         final long timeTo) {
-            return wrapped().fetchAll(timeFrom, timeTo);
-        }
-    }
-
-    private static class TimestampedWindowStoreReadOnlyDecorator<K, V>
-        extends WindowStoreReadOnlyDecorator<K, ValueAndTimestamp<V>>
-        implements TimestampedWindowStore<K, V> {
-
-        private TimestampedWindowStoreReadOnlyDecorator(final TimestampedWindowStore<K, V> inner) {
-            super(inner);
-        }
-    }
-
-    private static class SessionStoreReadOnlyDecorator<K, AGG>
-        extends StateStoreReadOnlyDecorator<SessionStore<K, AGG>, K, AGG>
-        implements SessionStore<K, AGG> {
-
-        private SessionStoreReadOnlyDecorator(final SessionStore<K, AGG> inner) {
-            super(inner);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, AGG> findSessions(final K key,
-                                                               final long earliestSessionEndTime,
-                                                               final long latestSessionStartTime) {
-            return wrapped().findSessions(key, earliestSessionEndTime, latestSessionStartTime);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, AGG> findSessions(final K keyFrom,
-                                                               final K keyTo,
-                                                               final long earliestSessionEndTime,
-                                                               final long latestSessionStartTime) {
-            return wrapped().findSessions(keyFrom, keyTo, earliestSessionEndTime, latestSessionStartTime);
-        }
-
-        @Override
-        public void remove(final Windowed<K> sessionKey) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public void put(final Windowed<K> sessionKey,
-                        final AGG aggregate) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public AGG fetchSession(final K key, final long startTime, final long endTime) {
-            return wrapped().fetchSession(key, startTime, endTime);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, AGG> fetch(final K key) {
-            return wrapped().fetch(key);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, AGG> fetch(final K from,
-                                                        final K to) {
-            return wrapped().fetch(from, to);
-        }
-    }
-
-    private abstract static class StateStoreReadWriteDecorator<T extends StateStore, K, V>
-        extends WrappedStateStore<T, K, V> {
-
-        static final String ERROR_MESSAGE = "This method may only be called by Kafka Streams";
-
-        private StateStoreReadWriteDecorator(final T inner) {
-            super(inner);
-        }
-
-        @Override
-        public void init(final ProcessorContext context,
-                         final StateStore root) {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-
-        @Override
-        public void close() {
-            throw new UnsupportedOperationException(ERROR_MESSAGE);
-        }
-    }
-
-    static class KeyValueStoreReadWriteDecorator<K, V>
-        extends StateStoreReadWriteDecorator<KeyValueStore<K, V>, K, V>
-        implements KeyValueStore<K, V> {
-
-        KeyValueStoreReadWriteDecorator(final KeyValueStore<K, V> inner) {
-            super(inner);
-        }
-
-        @Override
-        public V get(final K key) {
-            return wrapped().get(key);
-        }
-
-        @Override
-        public KeyValueIterator<K, V> range(final K from,
-                                            final K to) {
-            return wrapped().range(from, to);
-        }
-
-        @Override
-        public KeyValueIterator<K, V> all() {
-            return wrapped().all();
-        }
-
-        @Override
-        public long approximateNumEntries() {
-            return wrapped().approximateNumEntries();
-        }
-
-        @Override
-        public void put(final K key,
-                        final V value) {
-            wrapped().put(key, value);
-        }
-
-        @Override
-        public V putIfAbsent(final K key,
-                             final V value) {
-            return wrapped().putIfAbsent(key, value);
-        }
-
-        @Override
-        public void putAll(final List<KeyValue<K, V>> entries) {
-            wrapped().putAll(entries);
-        }
-
-        @Override
-        public V delete(final K key) {
-            return wrapped().delete(key);
-        }
-    }
-
-    static class TimestampedKeyValueStoreReadWriteDecorator<K, V>
-        extends KeyValueStoreReadWriteDecorator<K, ValueAndTimestamp<V>>
-        implements TimestampedKeyValueStore<K, V> {
-
-        TimestampedKeyValueStoreReadWriteDecorator(final TimestampedKeyValueStore<K, V> inner) {
-            super(inner);
-        }
-    }
-
-    static class WindowStoreReadWriteDecorator<K, V>
-        extends StateStoreReadWriteDecorator<WindowStore<K, V>, K, V>
-        implements WindowStore<K, V> {
-
-        WindowStoreReadWriteDecorator(final WindowStore<K, V> inner) {
-            super(inner);
-        }
-
-        @Deprecated
-        @Override
-        public void put(final K key,
-                        final V value) {
-            wrapped().put(key, value);
-        }
-
-        @Override
-        public void put(final K key,
-                        final V value,
-                        final long windowStartTimestamp) {
-            wrapped().put(key, value, windowStartTimestamp);
-        }
-
-        @Override
-        public V fetch(final K key,
-                       final long time) {
-            return wrapped().fetch(key, time);
-        }
-
-        @SuppressWarnings("deprecation") // note, this method must be kept if super#fetch(...) is removed
-        @Override
-        public WindowStoreIterator<V> fetch(final K key,
-                                            final long timeFrom,
-                                            final long timeTo) {
-            return wrapped().fetch(key, timeFrom, timeTo);
-        }
-
-        @SuppressWarnings("deprecation") // note, this method must be kept if super#fetch(...) is removed
-        @Override
-        public KeyValueIterator<Windowed<K>, V> fetch(final K from,
-                                                      final K to,
-                                                      final long timeFrom,
-                                                      final long timeTo) {
-            return wrapped().fetch(from, to, timeFrom, timeTo);
-        }
-
-        @SuppressWarnings("deprecation") // note, this method must be kept if super#fetch(...) is removed
-        @Override
-        public KeyValueIterator<Windowed<K>, V> fetchAll(final long timeFrom,
-                                                         final long timeTo) {
-            return wrapped().fetchAll(timeFrom, timeTo);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, V> all() {
-            return wrapped().all();
-        }
-    }
-
-    static class TimestampedWindowStoreReadWriteDecorator<K, V>
-        extends WindowStoreReadWriteDecorator<K, ValueAndTimestamp<V>>
-        implements TimestampedWindowStore<K, V> {
-
-        TimestampedWindowStoreReadWriteDecorator(final TimestampedWindowStore<K, V> inner) {
-            super(inner);
-        }
-    }
-
-    static class SessionStoreReadWriteDecorator<K, AGG>
-        extends StateStoreReadWriteDecorator<SessionStore<K, AGG>, K, AGG>
-        implements SessionStore<K, AGG> {
-
-        SessionStoreReadWriteDecorator(final SessionStore<K, AGG> inner) {
-            super(inner);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, AGG> findSessions(final K key,
-                                                               final long earliestSessionEndTime,
-                                                               final long latestSessionStartTime) {
-            return wrapped().findSessions(key, earliestSessionEndTime, latestSessionStartTime);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, AGG> findSessions(final K keyFrom,
-                                                               final K keyTo,
-                                                               final long earliestSessionEndTime,
-                                                               final long latestSessionStartTime) {
-            return wrapped().findSessions(keyFrom, keyTo, earliestSessionEndTime, latestSessionStartTime);
-        }
-
-        @Override
-        public void remove(final Windowed<K> sessionKey) {
-            wrapped().remove(sessionKey);
-        }
-
-        @Override
-        public void put(final Windowed<K> sessionKey,
-                        final AGG aggregate) {
-            wrapped().put(sessionKey, aggregate);
-        }
-
-        @Override
-        public AGG fetchSession(final K key,
-                                final long startTime,
-                                final long endTime) {
-            return wrapped().fetchSession(key, startTime, endTime);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, AGG> fetch(final K key) {
-            return wrapped().fetch(key);
-        }
-
-        @Override
-        public KeyValueIterator<Windowed<K>, AGG> fetch(final K from,
-                                                        final K to) {
-            return wrapped().fetch(from, to);
-        }
-    }
 }
